Update patch set 5
Patch Set 5: Verified; No score

Patch Set 5 is a clean cherry-pick of Patch Set 4 to avoid a future Gerrit merge conflict. Builds fine.

(In the future, it'd be helpful to repo sync the latest change and to rebase your local branch onto it, with "git rebase korg/master", as that helps Gerrit avoid conflicts. Thanks!)

Patch-set: 5