Update patch set 1
Patch Set 1: No score

It'd still be good to keep information across versions of a same change. E.g. in https://review.source.android.com/14726 you mentioned the bug being fixed, which is very relevant information for a reviewer and for anyone looking at the code in the future, but this information is gone from this change.

Patch-set: 1