Only display the Badge on Contacts, if there is data to display.
Just tap the Android symbol / contact Picture of a Contact with
no Phone number or Email. A empty Badge will open.
This fix prevents the Badge from opening, and immediately
finishes the calling Badge-Activity if there is no data to display
To test:
Create an empty Contact (only with a name), and click the Android symbol
on the Contact.
old Version: Empty Badge
new Version: no Badge

Change-Id: I8991a7db40b6df92d7f8ac43dfd23652f6ff23f5