Added correct OrientationChange handling
Dialogs in Contactedit are now handled correctly.
When you press Phonenumber Type and select Custom, a
Input Dialog pops up. There was an error when
changing Orientation (in example
open the Keyboard on the G1), the Dialog was gone
(with an internal WindowLeaked Exception). The new Manifest
entry handles it correctly, so the InputDialog will stay on top
even if Orientation is changed. All other fields also get populated
correctly, cause the Code to handle it is already there.

Change-Id: I242689846378a33408f62e825bb7755d1b93038d