Update patch set 1
Patch Set 1: No score

Is there another way in Android to handle Dialogs on Orientation change?
The only way i can think about is setting a boolean variable if a alertdialog is open, and call the .show() function again.

Maybe you can provide me a better solution for this.

Patch-set: 1