Update patch set 1
Patch Set 1:

Looks like this can not be done with a little code change, and i think a bigger change will not merge in the google branch. I will have a look at it once Froyo is released.
The Problem is, the Dialog is created and maintained in a own View Class that extends Relative/Absolute Layout, so i have no Activity to make this Dialog a managed one with setOwnerActivity.
Thx for your help anyway

Patch-set: 1