Update patch set 5
Patch Set 5: Fails

As it turns out, we've made a similar change in froyo, which sadly but fairly naturally conflicts with your change.

I can submit your change in the current open-source master but ignore its actual content for the purpose of merging it into Google's internal trees. This way you'll have credit for the change, even though your actual code won't survive in the long term. Let me know if that works for you. Otherwise we'll just consider abandoning this change.

Patch-set: 5