Update patch set 4
Patch Set 4:

Hi Dmitri,
first of all thank you for the effort you're putting in this change. I also thought about this "no badge behaviour" and how to avoid user bugs. I think a Text in the Badge is the simplest solution to do this job. I also thought about Toasts, but if the user clicks on the Symbol multiple times, this will result in a neverending toast-flash. Adding a Textview is also not so much work, since you only have to add a TextView instead of finishing the activity.

Another Question on this Badge thing:
I realy like this feature, most of all as a replacement for full screen dialogs if you want the user to choose what he wants. I was wondering why this feature is not implemented in the Framework so other apps can use it, and so i was thinking about a good way to put this in the framework.
Are there any ongoing changes on this in the internal Google branch on this, before i take a deeper look on this?

thanks
Christian

Patch-set: 4