Added correct OrientationChange handling
Dialogs for Custom Labels in Contactedit are now handled correctly.
When you press Phonenumber Type and select Custom, a
Input Dialog pops up. There was an error when
changing Orientation (in example
open the Keyboard on the G1), the Dialog was gone
(with an internal WindowLeaked Exception).
Now the Dialogstate, the EditText state and the Type for the Custom Label
are preserved by storing them in state.
The CustomLabel Dialog now also has a onDismissListener, so the private
variable will always be null when there is no Dialog to save ressources
and we do not leak a Context.

Change-Id: I77bdadc9c245e42895af04356848b2ac850b19cb