Update patch set 2
Patch Set 2:

In this case i can not use Managed Dialogs cause this all happens in a self written Relative Layout for the Contacts app. The only way to do this with Managed Dialogs would be to move lots of code to the Activity itself but in my optinion this would be too much work with too much side effects for this little change.

Patch-set: 2