Prevent crash when adding sync group
When adding a sync group the ContactsPreferencesActivity creates a URI for the
Contacts provider. The default Eclair (and Froyo) implemention does not
include the group ID which is required by the Contacts provider. This patch
appends the group ID to the URI path, which is where the Contacts provider
is looking for it.

Change-Id: Icd64d24141cd95b1d538e64d1ab4a04aac25fe76