Update patch set 1
Patch Set 1: Verified

Seems fine, merges and builds fine (both in open-source tree and Google's internal master). Tests pass.

Style-wise, I wonder whether there's be an opportunity to re-write that line in a slightly more readable way. It wasn't great before the change, and the change doesn't really improve it.

I'll let the primary maintainers comment on details, but I feel that this could be one of those places where the most readable code isn't the one that follows the style guide.

Patch-set: 1