Update patch set 1
Patch Set 1:

A correction on my part. The base of the problem is in the ContentUris class in the parseId method, as the ContactsProvider2 class calls this method to extract the group ID from the URI.

public static long parseId(Uri contentUri) {
    String last = contentUri.getLastPathSegment();
    return last == null ? -1 : Long.parseLong(last);
}

Is this the most appropriate place to make the correction?

Patch-set: 1