Update patch set 3
Patch Set 3:

I've created a new patch for this.  It should be uploaded shortly for review.

It changes much less and sticks to the original intended logic.  When I created this patch, I was a bit confused by the logic EntitySet was using, but now I believe I fully understand it.

Patch-set: 3