Call From SIM
When trying to call a number on the SIM, for example by dialing '1#',
Android code does not make a distinction between order and position
on the SIM. The contacts are sorted, which is wrong, and any empty
positions are removed (positions can be empty on a SIM). This means
that trying to make a phone call this way will lead to users dialing
the wrong number.

An API to load SIM positions has been added to solve the issue.

Change-Id: Ia6a921be75f0f969998745323acc60e0ae8d37e3