Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

I agree with Christian.  Since we aren't in an activity, we can't use managed dialogs.  This is not a big enough change and is a couple levels deep from the activity in order to try to move it to the activity.

Patch-set: 3