Update patch set 1
Patch Set 1:

The ContactsProvider2.java I originally noted this problem was the Eclair version. I tested the scenario on the vanilla Motorola Droid running Eclair and a Nexus One that had been upgraded to Froyo, and in both cases I saw the crash.

I did a file compare and did not see the effected line as you noted in the latest version. It appears that this bug has been corrected in Froyo. As I have seen this problem with on an upgraded Nexus One, is this perhaps a hiccup with the upgrade?

Patch-set: 1