Update patch set 1
Patch Set 1:

The referenced CL removes obtaining the group ID from the URI. The newer code appears to get the group ID from the selection arguments, which should be set a few lines below th modification in this patch. Is this correct?

Based on this correction being present in Froyo, this should only be a problem in Eclair devices. Since we upgraded our Nexus One from Eclair to Froyo, perhaps this is a problem with upgraded devices as well.

Patch-set: 1