Update patch set 1
Patch Set 1: Verified

The style looks fine, and the change merges/build fine in the master open-source tree.

However, I can't find any trace of that code in Google's internal master. It might have been re-written entire.

Daniel, does this ring a bell? It might make sense to merge it in the open-source tree if it looks good in that context, but then skip the merge entirely if that code doesn't make sense any more in our internal master tree.

Patch-set: 1