Update patch set 1
Patch Set 1:

Yeah, it is a timing thing, and you have to be perstent about it.

This is how the internal bug looks like:

Step 1. Tap ‘Home’ icon in the idle screen.
Step2. Tap ‘Contacts’ icon in the application tray.
Step3. Long Tap contact name(Gao) in the contact screen.
Step4. Tap the ‘Edit contact’ item in this popup screen.
Step5. Tap any one ‘+' icon continually(Such as tap ‘+' icon in the right of the ‘Organization’ item) about 4 times in the contact edit screen.
Step6. If you cannot reproduce this issue, please repeat step5.

I can reproduce this on my nexus one, but have to press the + about 20 times or so.

Patch-set: 1