Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

I couldn't reproduce the bug myself, but this seems like a safe patch that only improves things. Thanks for submitting!


JBQ: The equivalent code in master is LabeledEditorView.java (currently line 188). Should I patch this myself or is it better to do this as part of the auto-merge?

Patch-set: 1