Update patch set 1
Patch Set 1:

Daniel: thanks for pointing out the new location of the code. I found it.

Typically, we prefer to take care of that right as part of the conflict resolution, so that the history contains the best possible attribution for the origin of the change.

In this case, the conflict is actually simple enough that I'll have no problem dealing with it. Feel free to approve the change, and I'll submit it as soon as appropriate (i.e. as soon as I'm in a position to resolve the conflict immediately).

Patch-set: 1