Update patch set 3
Patch Set 3:

We have an internal bug tracking this issue.  Simply clearing the count is probably not a sufficient solution, because it is synced up and down with Google Contacts, is shared between applications including email and is used to order auto-complete suggestions among other things.  We should probably be looking for a comprehensive solution along the lines of "clear interaction history with this person".

Patch-set: 3