workaround for sometimes generated random call on clicked a call button in the call log
If a user releases a call button when the call log is redrawn
by the contacts information, sometimes wrong number is called
by a cached view with unorder.
So on clicking, a call number is used not a number in current cached view
but a number in a view at a touched moment
this is workaround.

Change-Id: I2d32f9177ab55f5df2bcf1c88d78a7e008609fcf
Signed-off-by: SungHyun Kwon <sh.kwon@lge.com>