Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Thank you for your CL and sorry for the long delay.

While this fixes the issue you described it also introduces incorrectly painted section headers (as described by the comment that would be removed by this CL). Not a big deal, but I would rather not trade in one tiny issue by another one.

This part of the code is undergoing bigger changes in ICS, so this change would have a limited effect in any case. Sorry, but I think this is currently not worth the trouble.

Patch-set: 1