Add lookup in Call logs to lookupNumber
https://android-review.googlesource.com/35390 added caller name display
to GSM phones. But whenever Call log is opened the information was overwritten,
if no contact is present with that phoen number.
This patch add a lookup function that tries to find a match in the Call log
(after other posibilities failed) and returns that information to be put in cache.

* Patch Set 2 : corrected coding style
* Patch Set 3 : - Query only incoming calls (the info we're looking for
                  if it exists is there and only there)
                - Correct phone number formatting

Change-Id: I0868d5f8341382639b5ebf9d5e616f118fc423c3