Update patch set 3
Patch Set 3: Do not submit

Looks like the other change has been merged?  I'm not familiar with the caller display name feature, but there's some implications of showing a non-contact name in calllog, and we're locking for a generic solution.  Until we find a generic solution, I think it'd be better not showing non-contact names on calllog.  Sorry.

Patch-set: 3
Label: Code-Review=-2