Update patch set 3
Patch Set 3:

in a perfect world the other patch is sufficient.
(the information is shown and kept as long as one doesn;t open the call log.)
once the call log is open the information is literally & explicitly overwritten in the cache b/c no contact is found,  right after lookupNumber is called. if that behavior is to be no more, this patch is not needed.


third party call log manager do no do this.

Patch-set: 3
Reviewer: Raouf M. Bencheraiet <1011783@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0