Do DataUsageFeedback update after starting ACTION_CALL
In QuickContactActivity, when clicking the Call button,
DataUsageFeedback update blocks the starting of ACTION_CALL
for 30-60ms.

"DataUsageFeedback update" and "ACTION_CALL" doesn't interfere with
each other. So it is harmless to update the DataUsageFeedback after
starting CallActivity.

Benefit:
Reduces 30-60ms of time needed to bring up the dialing UI.

Change-Id: I8a349b43c63f70724c89152a9e251f5a2230556f
Signed-off-by: LIANGTAO GAO <w20048@motorola.com>