Cannot open date before 1970/01/01 from quickcontact screen
The pre-install calendar support to jump the oldest date is 1970/01/01.
If date set before 1970/01/01, it can't jump to the date.

When save birthday or anniversary event for a contact,
set the year to 0, it means when click the birthday or anniversary
it will jump to the day of the coming birthday or anniversary.

Bug: Not yet
Test: Manual
Change-Id: I68d207cddbab53b97c1368048a50cda5d6bd5aa2