Fixed to Added the function to add the Phonetic name into Phonetic field automaticaly.
This fix is to add the Phonetic name automatically to Phonetic name field when user inputs a Name in Japanese language mode.

Enter in the name field and automatically enter to Phonetic field when the input charactor is confirmed.
The cases that will be entered to the Phonetic field are as follows.
1. Select enter on the keyboard
2. Select for prediction conversion candidate
3. Tap outside the input field.

Signed-off-by: yoichi kakimoto <youichi.kakimoto.gt@kyocera.jp>

(cherry picked from commit 9aa933cb743d7c88c3ef55ec2221e015f01ab923)

Change-Id: I2a1a89ce026ea3b3fab31f0c9f3859cf36cb49ad