Update patch set 1
Patch Set 1:

Thank you for review.

The design of the Contacts application is as described because it makes the unfixed character reflected in the name field reflect on the phonetic field.

The reason is that once confirmed predictive conversion name field string (kanji) and unfixed character to be reflected in the phonetic field can not be synchronized.
Example:
Name: kanji
Phonetic field (Furigana): hiragana

At the time the name field changes, the string is passed to addTextChangedListener ().

Contacts analyzes the character string delivered here and sets the unfixed character added to the phonetic field if it is an unfixed character.

The indoubt state is judged by whether or not Spanned.SPAN_COMPOSING is included in the span information acquired by getSpanFlags ().

In Google Japanese input, you enter in name field > Edit other field > Even if you focus on the name field again, the entered characters did not become uncertain characters but Gboard (Google's Keyboardad) "When you input a again, testa is in an undetermined state".

I think that what you pointed out depends on the difference of Google's Keyboad behavior.

If there is any way to avoid it please point out.

Patch-set: 1
Reviewer: yoichi kakimoto <1224854@85c56323-6fa9-3386-8a01-6480fb634889>