Update patch set 1
Patch Set 1: Code-Review-1

This commit introduces an issue within the SIM number and users handling.

In Settings / System / About / Status / SIM Status, the field "My phone number" will either be "Unknown" or the real number, depending on what is stored on the SIM card.

If it is the real number, the "Current user" in Settings dashboard will be listed as the number "2" instead of "Owner".

If it is "Unknown", the "Current user" is correctly shown as "Owner".
But if you open the AOSP Contacts app, edit your profile and add your phone number, the current user will shown as "2".

Patch-set: 1
Reviewer: Arne Coucheron <1032613@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1