Fix contact photo does not show issue
In Android Work Profile, the photo will not show in contact editor after
taking photo and saving it. This is because for Contacts in Android Work
Profile, PMS finds no handler for "com.android.camera.action.CROP" intent,
and the photo is too big without cropping to cache in ContactPhotoManager.

Solution is if the photo is too big to cache, try to compress the photo
into thumbnail size and cache it again, then cache it successfully.

Test: manual, ran the test 10 times and it passed 10 times
Bug: 121985416
Change-Id: Iad9012c438ec29f1c439a67ad90797435d748c6b